backfill: respect unread_count when marking unread
The unread_count property gives the number of messages that are unread
according to LinkedIn. This commit makes the bridge take that into
account when determining what to mark as read in Matrix.

This fixes a long-standing issue where we were totally not respecting
this field at all causing many annoyances with backfilling.